SMSC persistence update 02.04.2013

+++ core/slee/smpp-server-ra/ra/src/main/java/org/mobicents/smsc/slee/resources/smpp/server/SmppSessionsImpl.java
@@ -59,7 +59,7 @@ public class SmppSessionsImpl implements SmppSessions {
 
 	}
 
-	public String getEsmeClusterName(byte ton, byte npi, String address) {
+	public String getEsmeClusterName(int ton, int npi, String address) {
 		return this.smsRouteManagement.getEsmeClusterName(ton, npi, address);
 	}
